refactor(experiments): first pass to refactor the Metric Header into a reusable component.
